% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.density.LPJGUESS.R
\name{adjust.density.LPJGUESS}
\alias{adjust.density.LPJGUESS}
\title{Adjust LPJ-GUESS individual's density}
\usage{
adjust.density.LPJGUESS(individual, rel.change)
}
\arguments{
\item{individual}{A nested list which encapsulates an LPJ-GUESS 'Individual' as read from a binary state file}

\item{rel.change}{A numeric by which to scale the density and C and N pools}
}
\value{
the scaled 'individual' (the initial nested list with update values)
}
\description{
Very simple function that just scales the density of individuals and the associated C and N pools 
by a relative amount
}
\author{
Matthew Forrest
}
\keyword{internal}
