% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.LPJGUESS.R
\name{allometry}
\alias{allometry}
\title{LPJ-GUESS allometry}
\usage{
allometry(lifeform = "TREE", cmass_leaf, cmass_sap, cmass_heart,
  densindiv, age, fpc, deltafpc, sla, k_latosa, k_rp, k_allom1, k_allom2,
  k_allom3, wooddens, crownarea_max)
}
\description{
LPJ-GUESS allometry
}
\keyword{internal}
