% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGridcellVariablePerPFT.LPJGUESS.R
\name{calculateGridcellSOC}
\alias{calculateGridcellSOC}
\title{calculateGridcellSOC}
\usage{
calculateGridcellSOC(
  model.state,
  pools = c("SOILSTRUCT", "SOILMETA", "SOILMICRO", "SLOWSOM", "PASSIVESOM")
)
}
\arguments{
\item{model.state}{A large multiply-nested list containing the entire LPJ-GUESS state as read by 
function \code{readStateBinary.LPJGUESS}}

\item{pools}{"SOILSTRUCT","SOILMETA","SOILMICRO","SLOWSOM","PASSIVESOM" (see somdynam.cpp)}
}
\value{
The total SOC in Gridcell. The unit is consistent with cmass in state, usually kg C m-2
}
\description{
Calculates soil carbon
This should be put into the SDA procedure.
}
\author{
Yinghao Sun
}
