% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGridcellVariablePerPFT.LPJGUESS.R
\name{calculateGridcellVariablePerPFT}
\alias{calculateGridcellVariablePerPFT}
\title{calculateGridcellVariablePerPFT}
\usage{
calculateGridcellVariablePerPFT(model.state, variable)
}
\arguments{
\item{model.state}{A large multiply-nested list containing the entire LPJ-GUESS state as read by 
function \code{readStateBinary.LPJGUESS}}

\item{variable}{A character string specifying what variable to extract.  This can be chosen based on the LPJ-GUESS variable name
as recorded in the big list of list (that represents describes the model state in R).  Once special case is "biomass" which
returns the sum of "cmass_leaf", "cmass_root", "cmass_sap" and "cmass_heart"}
}
\value{
A numeric vector, with one entry per PFT
}
\description{
Calculates a per-PFT, gridcell-summed quantity from the LPJ-GUESS state, correctly averaging over patches.
This should be put into the SDA procedure.
}
\author{
Matthew Forrest
}
