% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_state.R
\name{extract_from_state_by_key}
\alias{extract_from_state_by_key}
\title{Extract nested value from a state list using flat key}
\usage{
extract_from_state_by_key(state, key)
}
\arguments{
\item{state}{A nested list (usually the model.state$state)}

\item{key}{A flat string like "Gridcell/Stand/1/Patch/1/Vegetation/Individuals/3/cmass_leaf"}
}
\value{
The value stored at that nested position
}
\description{
Extract nested value from a state list using flat key
}
\author{
Yinghao Sun
}
\keyword{internal}
