% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_closing}
\alias{find_closing}
\title{Find Closing Bracket}
\usage{
find_closing(find = "}", line_no, file_in, if_else_check = FALSE)
}
\arguments{
\item{find}{A character string of the opening bracket.}

\item{line_no}{A numeric value indicating the line number to start the search.}

\item{file_in}{A character vector of the file content.}

\item{if_else_check}{Optional. A logical value indicating whether to check for if/else blocks (default is FALSE).}
}
\value{
A numeric value indicating the line number of the matching closing bracket.
}
\description{
Identifies the line number of the matching closing bracket for a given opening bracket.
}
