% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_stream_size}
\alias{find_stream_size}
\title{Find Stream Size}
\usage{
find_stream_size(
  current_stream_type,
  guessh_in,
  LPJ_GUESS_TYPES,
  LPJ_GUESS_CONST_INTS
)
}
\arguments{
\item{current_stream_type}{A list containing details of the current stream.}

\item{guessh_in}{A character vector of LPJ-GUESS header file content.}

\item{LPJ_GUESS_TYPES}{A character vector of recognized LPJ-GUESS types.}

\item{LPJ_GUESS_CONST_INTS}{A dataframe mapping LPJ-GUESS constants to their values.}
}
\value{
A numeric value representing the size (number of streamed variables).
}
\description{
Determines the size (number of variables) in a stream based on the file content.
}
\keyword{internal}
