% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_stream_type}
\alias{find_stream_type}
\title{Find Stream Type}
\usage{
find_stream_type(
  class = NULL,
  current_stream_var,
  LPJ_GUESS_CLASSES,
  LPJ_GUESS_TYPES,
  guessh_in
)
}
\arguments{
\item{class}{A character string representing the class of the stream variable (default is NULL).}

\item{current_stream_var}{A character string representing the current stream variable.}

\item{LPJ_GUESS_CLASSES}{A character vector of LPJ-GUESS class names.}

\item{LPJ_GUESS_TYPES}{A character vector of recognized LPJ-GUESS types.}

\item{guessh_in}{A character vector of LPJ-GUESS header file content.}
}
\value{
A character string indicating the stream type.
}
\description{
Determines the type of a given stream variable in an LPJ-GUESS file.
}
\keyword{internal}
