% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_stream_var}
\alias{find_stream_var}
\title{Find Stream Variable}
\usage{
find_stream_var(file_in, line_nos)
}
\arguments{
\item{file_in}{A character vector representing the file content to search through.}

\item{line_nos}{A numeric vector of length 2, specifying the start and end lines to search for streamed variables.}
}
\value{
A character vector of streamed variable names.
}
\description{
A helper function that lists streamed variables. It returns the names of streamed variables.
}
\keyword{internal}
