% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model2netcdf.LPJGUESS}
\alias{model2netcdf.LPJGUESS}
\title{Function to convert LPJ-GUESS model output to standard netCDF format}
\usage{
model2netcdf.LPJGUESS(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of LPJ-GUESS model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert LPJ-GUESS output to netCDF
}
\author{
Istem Fer
}

