% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.LPJGUESS.R
\name{pecan2lpjguess}
\alias{pecan2lpjguess}
\title{Function to translate pecan param names and units to lpjguess names and units}
\usage{
pecan2lpjguess(trait.values)
}
\arguments{
\item{trait.values}{trait.values, list}
}
\value{
translated list
}
\description{
Function to translate pecan param names and units to lpjguess names and units
}
\author{
Istem Fer
}
