% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{read_binary_LPJGUESS}
\alias{read_binary_LPJGUESS}
\title{Read Binary File for LPJ-GUESS}
\usage{
read_binary_LPJGUESS(outdir, rundir, version = "PalEON")
}
\arguments{
\item{outdir}{The directory where the .state file is stored (e.g., .../state/1961)}

\item{rundir}{The directory where the corresponding params.ins is stored (e.g., .../run/ENS-00001)}

\item{version}{A character string specifying the LPJ-GUESS version (default is "PalEON").}
}
\value{
A matrix or list containing the extracted data.
}
\description{
Reads a binary file formatted for LPJ-GUESS and extracts relevant data.
}
\author{
Istem Fer, Yinghao Sun
}
