% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStateBinary.LPJGUESS.R
\name{readStateBinary}
\alias{readStateBinary}
\title{readStateBinary.LPJGUESS}
\usage{
readStateBinary(out.path, npft)
}
\arguments{
\item{out.path}{location on disk where model run outputs are stored}

\item{npft}{number of pfts specified in instruction file}
}
\value{
Patchpft_list state variables common to all individuals of a particular PFT
}
\description{
read binary state file of LPJ-GUESS
}
\author{
Istem Fer
}
