% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{serialize_starts_ends}
\alias{serialize_starts_ends}
\title{Serialize Starts and Ends}
\usage{
serialize_starts_ends(file_in, pattern = "void Gridcell::serialize")
}
\arguments{
\item{file_in}{A character vector of file lines to search through.}

\item{pattern}{A character string pattern to look for in the file.}
}
\value{
A numeric vector of length 2, giving the start and end line numbers.
}
\description{
Finds the start and end lines for serialization.
}
\keyword{internal}
