% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.state.LPJGUESS.R
\name{TotalCarbon}
\alias{TotalCarbon}
\title{Calculate Above Ground Woody Biomass}
\usage{
TotalCarbon(individual, include.debt = TRUE)
}
\description{
Calculate Above Ground Woody Biomass
}
\keyword{internal}
