% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateState.LPJGUESS.R
\name{updateState.LPJGUESS}
\alias{updateState.LPJGUESS}
\title{updateState.LPJGUESS}
\usage{
updateState.LPJGUESS(model.state, dens.initial, dens.target,
  biomass.initial, biomass.target)
}
\arguments{
\item{model.state}{A large multiply-nested list containing the entire LPJ-GUESS state as read by 
function \code{readStateBinary.LPJGUESS}}

\item{dens.initial}{A numeric vector of the initial stand-level stem densities (indiv/m^2) as named numeric vector 
with one entry per PFT/species, with the names being the PFT/species codes.  These values should be produced
using state data assimilation from function XXXXXX.}

\item{dens.target}{A numeric vector of the target stand-level stem densities (indiv/m^2) as named numeric vector 
with one entry per PFT/species, with the names being the PFT/species codes.  These values should be produced
using state data assimilation from function XXXXXX}

\item{biomass.target}{A numeric vector of the target stand-level biomasses (kgC/m^2) as named numeric vector 
with one entry per PFT/species, with the names being the PFT/species codes.  These values should be produced
using state data assimilation from function XXXXXX}

\item{biomass.target}{A numeric vector of the target stand-level biomasses (kgC/m^2) as named numeric vector 
with one entry per PFT/species, with the names being the PFT/species codes.  These values should be produced
using state data assimilation from function XXXXXX}
}
\value{
And updated model state (as a big old list o' lists)
}
\description{

}
\details{
Adjust LPJ-GUESS state
}
\author{
Matthew Forrest
}
