% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.config.LPJGUESS}
\alias{write.config.LPJGUESS}
\title{Write LPJ-GUESS configuration files}
\usage{
write.config.LPJGUESS(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}

\item{trait.samples}{vector of samples for a given trait}
}
\value{
configuration file for LPJ-GUESS for given run
}
\description{
Writes a LPJ-GUESS config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Istem Fer, Tony Gardella
}

