% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.LPJGUESS.R
\name{write.insfile.LPJGUESS}
\alias{write.insfile.LPJGUESS}
\title{Write LPJ-GUESS instruction script}
\usage{
write.insfile.LPJGUESS(settings, trait.values, rundir, outdir, run.id)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{run.id}{PEcAn run ID}
}
\value{
settings Updated list
}
\author{
Istem Fer
}

