% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_state.R
\name{write_binary_LPJGUESS}
\alias{write_binary_LPJGUESS}
\title{Write updated variables into a copy of the original LPJ-GUESS .state file}
\usage{
write_binary_LPJGUESS(State_updated, outdir)
}
\arguments{
\item{State_updated}{A list containing updated state variables, position list and size list (get from read_binary)}

\item{outdir}{Path to a directory containing the `0.state` and `meta.bin` files.}
}
\value{
No return value. Writes files to disk as side effect.
}
\description{
Write updated variables into a copy of the original LPJ-GUESS .state file
}
\author{
Yinghao Sun
}
