% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.state.LPJGUESS.R
\name{AbvGrndWood}
\alias{AbvGrndWood}
\title{Calculate Above Ground Woody Biomass}
\usage{
AbvGrndWood(individual, include.debt = TRUE)
}
\description{
Calculate Above Ground Woody Biomass
}
\keyword{internal}
