% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.biomass.LPJGUESS.R
\name{adjust.biomass.LPJGUESS}
\alias{adjust.biomass.LPJGUESS}
\title{Adjust LPJ-GUESS individual's biomass}
\usage{
adjust.biomass.LPJGUESS(individual, rel.change, sla, wooddens, lifeform,
  k_latosa, k_allom2, k_allom3)
}
\arguments{
\item{individual}{A nested list which encapsulates an LPJ-GUESS 'Individual' as read from a binary state file}

\item{rel.change}{A numeric by which to scale the density and C and N pools}

\item{sla}{The SLA (specific leaf area) (per PFT parameter)}

\item{wooddens}{Wood density (kgC/m^2) (per PFT parameter)}

\item{lifeform}{An integer code for the lifeform of this individual (cohort): 1 = Tree, 2 = Grass

The changes in C pools are determined by the allocation.  The changes in the N pools are designed to 
maintain the pre-exisiing C:N ratios, so N is just scaled using the updated C with the initial C:N ratio.
The N storage pools (nstore_longterm and nstore_labile) don't have pre-existing C:N ratios, so they are 
just scaled by the overall biomass change (the 'rel.change' argument to the function).

Note that after this function is called the function \code{allometry} should be used to update the individual
and to check that the newly updated individual has a 'valid' allometry. The litter pools should also be updated.
This is implemented in the \code{updateState} function following the call to this \code{adjustBiomass} function.}

\item{k_latosa}{The leaf area to sapwood area ratio (per PFT parameter)}

\item{k_allom2, k_allom3, }{Allometry coefficients (per PFT parameters)}

\item{crownarea_max}{Maximum allowed crown area (m^2)  (per PFT parameter)}
}
\value{
the scaled 'individual' (the initial nested list with update values)
}
\description{
This function adjusts an LPJ-GUESS individual by calling the LPJ-GUESS allocation function (compiled C++)
with a given biomass change.  It updates the individual biomass pools directly, and also returns, in a list further 
adjustments to the litter pools.
}
\author{
Matthew Forrest
}
\keyword{internal}
