% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_stream_size}
\alias{find_stream_size}
\title{Find Stream Size}
\usage{
find_stream_size(
  current_stream_type,
  guessh_in,
  LPJ_GUESS_TYPES,
  LPJ_GUESS_CONST_INTS
)
}
\arguments{
\item{file_in}{A character vector representing the file content.}

\item{line_nos}{A numeric vector of length 2, specifying the start and end lines to check for streamed variables.}
}
\value{
A numeric value representing the size (number of streamed variables).
}
\description{
Determines the size (number of variables) in a stream based on the file content.
}
