% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{find_stream_type}
\alias{find_stream_type}
\title{Find Stream Type}
\usage{
find_stream_type(
  class = NULL,
  current_stream_var,
  LPJ_GUESS_CLASSES,
  LPJ_GUESS_TYPES,
  guessh_in
)
}
\arguments{
\item{class}{A character string representing the class of the current stream, default is NULL.}

\item{LPJ_GUESS_CLASSES}{A list containing mappings of classes to stream types.}

\item{current_stream_type}{A character string representing the current stream type, default is NULL.}
}
\value{
A character string indicating the stream type.
}
\description{
Determines the type of stream variables in the file content.
}
\examples{
find_stream_type(class = "soil", current_stream_type = NULL, LPJ_GUESS_CLASSES = list(soil = "stream"))
}
