% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{read_binary_LPJGUESS}
\alias{read_binary_LPJGUESS}
\title{Read Binary File for LPJ-GUESS}
\usage{
read_binary_LPJGUESS(outdir, version = "PalEON")
}
\arguments{
\item{file_path}{A character string specifying the path to the binary file.}

\item{variable_name}{A character string specifying the name of the variable to extract.}
}
\value{
A matrix or list containing the extracted data.
}
\description{
Reads a binary file formatted for LPJ-GUESS and converts it into a usable structure.
}
