% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.LPJGUESS.R
\name{read_restart.LPJGUESS}
\alias{read_restart.LPJGUESS}
\title{Read Restart for LPJGUESS}
\usage{
read_restart.LPJGUESS(outdir, runid, stop.time, settings, var.names, params)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{var.names to be extracted}

\item{params}{passed on to return value}
}
\value{
X_tmp      vector of forecasts
}
\description{
Read Restart for LPJGUESS
}
\examples{
# example code
outdir = "/fs/data2/output//PEcAn_1000010473/out"
runid = 1002656839
stop.time = "1960-12-31 23:59:59 UTC"
load("/fs/data2/output/PEcAn_1000010473/SDAsettings_develop.Rdata")
var.names = c("AGB.pft", "TotSoilCarb")
load("/fs/data2/output/PEcAn_1000010473/SDAparams_develop.Rdata")
}
\author{
Istem Fer, Yinghao Sun
}
